Mount multiple overlayfs trees within the sandbox
After the sundering, the shattered remnants of the world were cast into
disarray.  Brother turned against brother, friend against friend, and
unlikely alliances were made with the kernel so as to preserve the
immutability of build image shards.  In the end however, the powers that
be determined that a mutable, yet evanescent landscape was necessary in
order that proper binaries might grow and flourish in the universe they
were constructing.  And so, a bold plan was devised, to mount an overlay
filesystem for each shard of the formerly whole build image.