Ensure `~/.gitconfig` has the information we need to make commits
If we do not have the information needed in a global `~/.gitconfig`,
then ask the user for it during wizard build time, and store that into a
repository-local `.git/config` file.  Also tell the user how to
generate a global `~/.gitconfig` file so that we don't ask this again,
and set that up as a `LibGit2` operation so we still don't need to shell
out to `git`.