Insulate `dlopen()` during auditing
When loading `.so` files that use LLVM, oftentimes internal LLVM
functions are called within the initialization routines of the `.so`
files.  This can cause Julia to segfault, which is bad.  Let's allow it
to continue segfaulting, but do so in an insulated Julia process so that
we don't end up killing ourselves during our audit phase.