Rely completely on `RPATH` to find libraries on nixy systems
Given that we know we _need_ to override paths via `LD_LIBRARY_PATH`/
`DYLD_FALLBACK_LIBRARY_PATH`, we nevertheless want to maintain as much
usability as possible without those environment variables set.  To do
so, we encode into the `RPATH` the relative location of the libraries we
want to link against, and then we set up our `DT_NEEDED` entries to be
as simple as possible: either the basename of the library, or
`@rpath/<basename>`, as the case may be on OSX.

This is the only combination that I have found allowing both native
execution under ideal conditions, AND overriding through environment
variables.