docker workaround MSYS path conversion issue
Without this on a Windows system with Git for Windows installed, I get:
```
docker: Error response from daemon: the working directory '\workspace\' is invalid, it needs to be an absolute path.
```

Apparently this is a known issue with Git for Windows: https://github.com/docker/for-win/issues/1971#issuecomment-383826853

It seems harmless to prepend a slash on any platform, to work around this.