Remove audit test that would fail on some systems
The audit test on `other_platform` assumes that there is a single libgfortran
available in the system, but this may not be true, because Julia ships its own
libgfortran that may not match the system one.

For example, if

```julia
our_libgfortran_version = v"4.0.0"
other_libgfortran_version = v"5.0.0"
```

and the system has actually libgfortran v5, trying to run `hello_world` built
for `other_libgfortran_version` would succeed because the dynamic loader would
use the system libgfortran, but the test is expecting a failure.

This commit simply removes this test, as there is no simple and general way to
make it reliably work on all platforms.