Dereference symlinks when copying `DirectorySource`s
Usually, we want to preserve symlinks; but in the special case of a `DirectorySource()`, it is very useful to be able to have symlinks pointing out side of the current directory and at some shared patch pool (e.g. in `GCCBootstrap@X`).  So we make the compromise that symlinks are collapsed when including `DirectorySource`'s.