Remove unnecessary duplicates in LIBPATH_list
We don't need to pass so many duplicates of Julia's libdir up through
the `LIBPATH_list` machinery in JLL packages; we can add it once at the
end for each package since we know this will always be the same for all
JLL packages.

This has the added bonus of fixing the Git_jll issues on MacOS, where
having Julia's libpcre in front of the installed LibPCRE2_jll was
causign problems.