Add optional init_block kwarg to build_tarball
This optional argument can be set to a string, which is then inserted at
the end of the generated __init__ methods. This allows inserting code
which initializes the JLL or in other ways makes it easier to use; e.g
for ncurses, it could set the TERMINFO env variable; for GMP, it could call
the `mp_set_memory_functions` function to ensure this copy of GMP uses
the same alloc/free functions as the GMP bundled with Julia.