Improve local development workflow (#861)
* Improve local development workflow

This adds two new capabilities:

* `--deploy=local`; for those who really just want to test out their new
JLLs locally, this makes it a cinch.  Just tack `--deploy=local` onto
the end of your `build_tarballs.jl` invocation, and you'll get the
artifacts installed and the JLL packages built within your
`~/.julia/dev` directory. (Note that you will still need to add that
directory into any environment you want to use it from)

* `override` directories for `dev`'ed out JLL packages.  This allows a
user to override the binaries within a JLL easily and cheaply.  At
compile time, the JLL package will look for an `override` directory in
its project root, and if it finds it, it will attempt to load binaries
from there instead of from an Artifact.  To aid in setting this up, JLLs
now have a `dev_jll()` method that will perform the checkout and copy
over the normal artifact's contents to the `override` directory,
allowing for the user to easily see what must be overwritten in order to
test out a local build of something.

* Tweak docs