Add support for parallel universes
This adds the concept of a "universe" for autobuilds. A universe
is essentially a depot that has its own copy of the General registry.
When building packages using `--deploy=universe` the built jll gets
registered into this copy of the registry and subsequent builds can
pick it up from there as if it had been registered with the real
general registry. This allows successive builds of dependent packages
without first having to go register the dependency. The universe
to use is specified using `--universe=<name>`. If a universe by
that name does not exist, it is created. Otherwise the previously
created universe is re-used.