Allow disabling the `isidentifier` check in `build_tarballs`
Validation of the field that will determine the name of the generated
JLL package is what you want 99.9% of the time. But that other 0.1% of
the time is `Yggdrasil/0_RootFS/PlatformSupport/build_tarballs.jl`: each
invocation of the script specifies a separate platform, the triplet of
which gets embedded in the name. Unfortunately e.g.
`PlatformSupport-x86_64-unknown-freebsd13.2` is not a valid identifier,
so building PlatformSupport with recent BinaryBuilder fails.