Remove separate source_urls wizard field
Goes with https://github.com/JuliaPackaging/BinaryBuilderBase.jl/pull/322.
Part of the problem with the current setup is that the WizardState doesn't
actually round trip properly, because the cache for archive downloads gets
deleted. It could be reconstituted from source_urls of course, but it would
be cleaner to just do that on-demand with the SetupSource (though this
commit doesn't do that yet).