Fix `with_libgfortran()` with multiple libpath directories
The creation of `libpath_list` assumed that there would only be one directory in
`LD_LIBRARY_PATH` (or the equivalents on other platforms), if there were more
then creating the list would fail because Julia would try to `hcat()` all the
elements together and throw an error about them having different
lengths.

Somewhat scarily, this would've silently done the wrong thing if `libpath_list`
was a column vector instead of a row vector O_o (i.e. if it originally had a
comma between the elements)