Bump minor version for a breaking release
As of this version, the package will error upon loading if used with
Julia > 1.7. Previously it could be loaded but didn't work properly with
later Julia versions. To be safe, we're making this a breaking release.
It should be a largely user-invisible change though, as BinaryBuilder is
generally not used as a package dependency.