Updates for FreeBSD AArch64 support (#1342)
* Updates for FreeBSD AArch64 support

Changes:
- Include FreeBSD AArch64 in minimum Julia compat check as is done for
  macOS AArch64.
- Mention FreeBSD AArch64 in the documentation for what's built by
  default.
- Avoid installing the Rust compiler shards on FreeBSD AArch64 because
  they don't exist (yet?!?!).

* Delete and ignore top-level Manifest.toml

By request

* Bump minor version for a breaking release

As of this version, the package will error upon loading if used with
Julia > 1.7. Previously it could be loaded but didn't work properly with
later Julia versions. To be safe, we're making this a breaking release.
It should be a largely user-invisible change though, as BinaryBuilder is
generally not used as a package dependency.